Tracking order

+++ lib/model/bao_cao_san_xuat_response.dart
@@ -10,7 +10,7 @@ class BaoCaoSanXuatResponse {
 
 class BaoCaoSanXuat {
   late String company;
-  late double value;
+  late int value;
 
   BaoCaoSanXuat.fromJson(Map<String, dynamic> json) {
     company = json['công ty'];
